Move render command to pkg/cmd
